move grids and nroots
