+++ triples/core/structsos/utils.py
@@ -5,7 +5,7 @@ import sympy as sp
 from ..shared import uniquely_named_symbol, SS
 from ...sdp import congruence as _congruence
 from ...utils.expression import Coeff, CyclicSum, CyclicProduct
-from ...utils.roots.rationalize import nroots, rationalize_bound
+from ...utils.roots import nroots, rationalize_bound
 
 class StructuralSOSError(Exception): ...
