+++ triples/gui/__init__.py
@@ -1,10 +1,12 @@
 from .sos_manager import SOS_Manager
 
+from .grid import GridPoly, GridRender
+
 from .visualize import (
     show_dets, plot_contour, plot_f
 )
 
 __all__ = [
-    'SOS_Manager',
+    'SOS_Manager', 'GridPoly', 'GridRender',
     'show_dets', 'plot_contour', 'plot_f'
 ]
